test(card): 카드 관련 도메인 테스트 정리
작업 내용
- 컨밴션 수정
- 패키지 정리
